// web/src/lib/api/schema.gen.ts
export interface components {
    schemas: {
// ...
        UpdateShoppingItemRequest: {
            version?: number;
            name?: string;
            /** Format: double */
            quantity?: number | null;
            /** @enum {string|null} */
            unit?: "g" | "ml" | "piece" | null;
            category?: components["schemas"]["IngredientCategory"];
            checked?: boolean;
        };
// ...
    };
// ...
}